fix(altera): solve s10 warm reset issue.
This is extension of the patch to fix the warm
reset issue in agilex refer to this commit 7f4fa931a.

The warm reset not able to trigger due to the system
not able to detect the magic number. ATF only able
to solve for boot core. For secondary cores, Linux
need to update psci driver to WFI the cores in EL3.
Original Linux WFI is EL1. Thus causing secondary cores not working

Signed-off-by: Boon Khai Ng <boon.khai.ng@altera.com>